import export with interactive list
