+++ .eslintrc.json
@@ -11,6 +11,7 @@
     },
     "extends": ["airbnb-base"],
     "rules": {
+      "import/prefer-default-export" : "off",
       "no-use-before-define": "off",
       "no-plusplus": "off",
       "max-classes-per-file": "off",
