+++ main.js
@@ -18,7 +18,7 @@
   \*************************************************************/
 /***/ ((module, __webpack_exports__, __webpack_require__) => {
 
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var _node_modules_css_loader_dist_runtime_noSourceMaps_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../node_modules/css-loader/dist/runtime/noSourceMaps.js */ \"./node_modules/css-loader/dist/runtime/noSourceMaps.js\");\n/* harmony import */ var _node_modules_css_loader_dist_runtime_noSourceMaps_js__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(_node_modules_css_loader_dist_runtime_noSourceMaps_js__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var _node_modules_css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../node_modules/css-loader/dist/runtime/api.js */ \"./node_modules/css-loader/dist/runtime/api.js\");\n/* harmony import */ var _node_modules_css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(_node_modules_css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_1__);\n// Imports\n\n\nvar ___CSS_LOADER_EXPORT___ = _node_modules_css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_1___default()((_node_modules_css_loader_dist_runtime_noSourceMaps_js__WEBPACK_IMPORTED_MODULE_0___default()));\n// Module\n___CSS_LOADER_EXPORT___.push([module.id, \"body {\\n  width: 100%;\\n  background-color: antiquewhite;\\n  margin: 0;\\n}\\n\\nmain {\\n  width: 100%;\\n  margin: 50px auto;\\n}\\n\\ni {\\n  cursor: pointer;\\n  color: gray;\\n}\\n\\n.to-do-list {\\n  display: flex;\\n  flex-flow: column;\\n  list-style-type: none;\\n  background-color: white;\\n  margin: 100px;\\n  padding: 0;\\n}\\n\\n.to-do-list > * {\\n  display: flex;\\n  border-bottom: 1px solid gray;\\n  flex: 1;\\n  padding: 16px;\\n}\\n\\n.title {\\n  justify-content: space-between;\\n  align-items: center;\\n}\\n\\n.title h3 {\\n  margin: 0;\\n}\\n\\n#form {\\n  width: 100%;\\n  display: flex;\\n  justify-content: space-around;\\n  align-items: center;\\n}\\n\\n#form input {\\n  border: 1px hidden;\\n  flex-basis: 90%;\\n  outline: none;\\n  font-size: 1rem;\\n  color: black;\\n}\\n\\n#form input[placeholder] {\\n  font-style: italic;\\n}\\n\\n#form button {\\n  border: 1px hidden;\\n  background-color: white;\\n  cursor: pointer;\\n}\\n\\n#form i {\\n  transform: rotate(0.5turn);\\n}\\n\\n.list-section {\\n  display: none;\\n  padding: 0;\\n}\\n\\n.allTodos {\\n  padding: 0;\\n}\\n\\n.todo {\\n  display: flex;\\n  justify-content: space-around;\\n  align-items: center;\\n  border-bottom: 1px solid gray;\\n  flex: 1;\\n  padding: 1rem;\\n}\\n\\n.todo:hover {\\n  cursor: pointer;\\n  background-color: whitesmoke;\\n}\\n\\n.todo div {\\n  flex-basis: 90%;\\n  display: flex;\\n  flex-direction: row;\\n  align-items: center;\\n  gap: 0.5rem;\\n}\\n\\n.todo p {\\n  flex-basis: 95%;\\n  margin: 0;\\n  color: black;\\n  font-size: 1.2rem;\\n  outline: none;\\n}\\n\\n#checkbox {\\n  width: 20px;\\n  height: 20px;\\n}\\n\\n#clear-completed {\\n  background-color: whitesmoke;\\n  justify-content: center;\\n  cursor: pointer;\\n}\\n\\n#clear {\\n  border: 1px hidden;\\n  background-color: ghostwhite;\\n  text-align: center;\\n  font-size: 1rem;\\n  cursor: pointer;\\n}\\n\\n.edit {\\n  margin-right: 10px;\\n}\", \"\"]);\n// Exports\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (___CSS_LOADER_EXPORT___);\n\n\n//# sourceURL=webpack://webpack-template/./src/style.css?./node_modules/css-loader/dist/cjs.js");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var _node_modules_css_loader_dist_runtime_noSourceMaps_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../node_modules/css-loader/dist/runtime/noSourceMaps.js */ \"./node_modules/css-loader/dist/runtime/noSourceMaps.js\");\n/* harmony import */ var _node_modules_css_loader_dist_runtime_noSourceMaps_js__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(_node_modules_css_loader_dist_runtime_noSourceMaps_js__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var _node_modules_css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../node_modules/css-loader/dist/runtime/api.js */ \"./node_modules/css-loader/dist/runtime/api.js\");\n/* harmony import */ var _node_modules_css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(_node_modules_css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_1__);\n// Imports\n\n\nvar ___CSS_LOADER_EXPORT___ = _node_modules_css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_1___default()((_node_modules_css_loader_dist_runtime_noSourceMaps_js__WEBPACK_IMPORTED_MODULE_0___default()));\n// Module\n___CSS_LOADER_EXPORT___.push([module.id, \"body {\\n  width: 100%;\\n  background-color: antiquewhite;\\n  margin: 0;\\n}\\n\\nmain {\\n  width: 100%;\\n  margin: 50px auto;\\n}\\n\\ni {\\n  cursor: pointer;\\n  color: gray;\\n}\\n\\n.to-do-list {\\n  display: flex;\\n  flex-flow: column;\\n  list-style-type: none;\\n  background-color: white;\\n  margin: 100px;\\n  padding: 0;\\n}\\n\\n.to-do-list > * {\\n  display: flex;\\n  border-bottom: 1px solid gray;\\n  flex: 1;\\n  padding: 16px;\\n}\\n\\n.title {\\n  justify-content: space-between;\\n  align-items: center;\\n}\\n\\n.title h3 {\\n  margin: 0;\\n}\\n\\n#form {\\n  width: 100%;\\n  display: flex;\\n  justify-content: space-around;\\n  align-items: center;\\n}\\n\\n#form input {\\n  border: 1px hidden;\\n  flex-basis: 90%;\\n  outline: none;\\n  font-size: 1rem;\\n  color: black;\\n}\\n\\n#form input[placeholder] {\\n  font-style: italic;\\n}\\n\\n#form button {\\n  border: 1px hidden;\\n  background-color: white;\\n  cursor: pointer;\\n}\\n\\n#form i {\\n  transform: rotate(0.5turn);\\n}\\n\\n.list-section {\\n  display: none;\\n  padding: 0;\\n}\\n\\n.allTodos {\\n  padding: 0;\\n}\\n\\n.todo {\\n  display: flex;\\n  justify-content: space-around;\\n  align-items: center;\\n  border-bottom: 1px solid gray;\\n  flex: 1;\\n  padding: 1rem;\\n}\\n\\n.todo:hover {\\n  cursor: pointer;\\n  background-color: whitesmoke;\\n}\\n\\n.todo div {\\n  flex-basis: 90%;\\n  display: flex;\\n  flex-direction: row;\\n  align-items: center;\\n  gap: 0.5rem;\\n}\\n\\n.todo p {\\n  flex-basis: 95%;\\n  margin: 0;\\n  color: black;\\n  font-size: 1.2rem;\\n  outline: none;\\n}\\n\\n#checkbox {\\n  width: 20px;\\n  height: 20px;\\n}\\n\\n#clear-completed {\\n  background-color: whitesmoke;\\n  justify-content: center;\\n  cursor: pointer;\\n}\\n\\n#clear {\\n  border: 1px hidden;\\n  background-color: ghostwhite;\\n  text-align: center;\\n  font-size: 1rem;\\n  cursor: pointer;\\n}\\n\\n.edit {\\n  margin-right: 10px;\\n}\\n\", \"\"]);\n// Exports\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (___CSS_LOADER_EXPORT___);\n\n\n//# sourceURL=webpack://webpack-template/./src/style.css?./node_modules/css-loader/dist/cjs.js");
 
 /***/ }),
 
@@ -118,7 +118,57 @@ eval("\n\n/* istanbul ignore next  */\nfunction styleTagTransform(css, styleElem
   \**********************/
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _style_css__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./style.css */ \"./src/style.css\");\n\n\nclass TaskList {\n  todos = [];\n\n  save() {\n    const todos = JSON.stringify(this.todos);\n    localStorage.setItem('todos', todos);\n  }\n\n  add(description) {\n    const todo = {\n      description,\n      completed: false,\n      index: this.todos.length + 1,\n    };\n    this.todos.push(todo);\n    this.save();\n  }\n\n  delete(id) {\n    this.todos.splice(id - 1, 1);\n    this.todos.forEach((i) => {\n      if (i.index > id) {\n        i.index -= 1;\n      }\n    });\n    this.save();\n  }\n\n  getReserved() {\n    this.todos = JSON.parse(localStorage.getItem('todos'));\n  }\n\n  edit(value, index) {\n    this.todos[index - 1].description = value;\n    this.save();\n  }\n\n  completed(status, index) {\n    this.todos[index - 1].completed = status;\n    this.save();\n  }\n\n  clearCompleted() {\n    this.todos = this.todos.filter((a) => a.completed === false);\n    for (let i = 0; i < this.todos.length; i += 1) {\n      this.todos[i].index = i + 1;\n    }\n    this.save();\n  }\n}\n\nconst todo = new TaskList();\nconst listSection = document.querySelector('.list-section');\n\nconst create = () => {\n  listSection.replaceChildren();\n  if (todo.todos.length > 0) {\n    listSection.style.display = 'block';\n    const listContainer = document.createElement('ul');\n    listContainer.className = 'allTodos';\n    listSection.appendChild(listContainer);\n    todo.todos.map((a) => {\n      const list = document.createElement('li');\n      list.className = 'todo';\n\n      const descriptionContainer = document.createElement('div');\n      const checkbox = document.createElement('input');\n      checkbox.type = 'checkbox';\n      checkbox.id = 'checkbox';\n      if (a.completed === true) {\n        checkbox.checked = 'checked';\n      }\n\n      checkbox.onclick = (e) => {\n        todo.completed(e.target.checked, a.index);\n      };\n\n      descriptionContainer.appendChild(checkbox);\n\n      const descript = document.createElement('p');\n      descript.id = 'task-description';\n      descript.textContent = a.description;\n      descriptionContainer.appendChild(descript);\n      list.appendChild(descriptionContainer);\n\n      const editIcon = document.createElement('i');\n      editIcon.className = 'fa fa-solid fa-pen-to-square';\n      list.appendChild(editIcon);\n\n      const deleteIcon = document.createElement('i');\n      deleteIcon.className = 'fa fa-solid fa-trash';\n      deleteIcon.id = a.index;\n\n      list.onclick = () => {\n        descript.contentEditable = 'true';\n        list.style.backgroundColor = 'whitesmoke';\n        list.appendChild(deleteIcon);\n        editIcon.style.display = 'none';\n        list.addEventListener('keydown', () => {\n          todo.edit(descript.innerHTML, a.index);\n        });\n      };\n      deleteIcon.onclick = () => {\n        todo.delete(a.index);\n        todo.save();\n        create();\n      };\n      listContainer.append(list);\n      return list;\n    });\n    listSection.appendChild(listContainer);\n  }\n};\n\nconst form = document.getElementById('form');\n\nconst getAdded = () => {\n  const description = document.getElementById('new-todo').value;\n  if (description != '') {\n    todo.add(description);\n    create();\n    form.reset();\n  }\n};\n\nconst populate = () => {\n  if (localStorage.getItem('todos')) {\n    todo.getReserved();\n    create();\n  } else {\n    create();\n  }\n};\n\nconst clearCompletedBtn = document.getElementById('clear-completed');\nconst clearCompleted = () => {\n  clearCompletedBtn.addEventListener('click', () => {\n    todo.clearCompleted();\n    create();\n  });\n};\n\npopulate();\nform.addEventListener('submit', getAdded);\nclearCompleted();\n\n//# sourceURL=webpack://webpack-template/./src/index.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _style_css__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./style.css */ \"./src/style.css\");\n/* harmony import */ var _modules_addTodo_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./modules/addTodo.js */ \"./src/modules/addTodo.js\");\n/* harmony import */ var _modules_populate_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./modules/populate.js */ \"./src/modules/populate.js\");\n/* harmony import */ var _modules_clear_js__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./modules/clear.js */ \"./src/modules/clear.js\");\n\n\n\n\n\n\n(0,_modules_populate_js__WEBPACK_IMPORTED_MODULE_2__.populate)();\n_modules_addTodo_js__WEBPACK_IMPORTED_MODULE_1__.form.addEventListener('submit', _modules_addTodo_js__WEBPACK_IMPORTED_MODULE_1__.getAdded);\n(0,_modules_clear_js__WEBPACK_IMPORTED_MODULE_3__.clearCompleted)();\n\n//# sourceURL=webpack://webpack-template/./src/index.js?");
+
+/***/ }),
+
+/***/ "./src/modules/addTodo.js":
+/*!********************************!*\
+  !*** ./src/modules/addTodo.js ***!
+  \********************************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
+
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"form\": () => (/* binding */ form),\n/* harmony export */   \"getAdded\": () => (/* binding */ getAdded)\n/* harmony export */ });\n/* harmony import */ var _showList_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./showList.js */ \"./src/modules/showList.js\");\n\n\nconst form = document.getElementById('form');\n\nconst getAdded = () => {\n  const description = document.getElementById('new-todo').value;\n  if (description != '') {\n    _showList_js__WEBPACK_IMPORTED_MODULE_0__.todo.add(description);\n    (0,_showList_js__WEBPACK_IMPORTED_MODULE_0__.create)();\n    form.reset();\n  }\n};\n\n//# sourceURL=webpack://webpack-template/./src/modules/addTodo.js?");
+
+/***/ }),
+
+/***/ "./src/modules/classList.js":
+/*!**********************************!*\
+  !*** ./src/modules/classList.js ***!
+  \**********************************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
+
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"TaskList\": () => (/* binding */ TaskList)\n/* harmony export */ });\nclass TaskList {\n    todos = [];\n  \n    save() {\n      const todos = JSON.stringify(this.todos);\n      localStorage.setItem('todos', todos);\n    }\n  \n    add(description) {\n      const todo = {\n        description,\n        completed: false,\n        index: this.todos.length + 1,\n      };\n      this.todos.push(todo);\n      this.save();\n    }\n  \n    delete(id) {\n      this.todos.splice(id - 1, 1);\n      this.todos.forEach((i) => {\n        if (i.index > id) {\n          i.index -= 1;\n        }\n      });\n      this.save();\n    }\n  \n    getReserved() {\n      this.todos = JSON.parse(localStorage.getItem('todos'));\n    }\n  \n    edit(value, index) {\n      this.todos[index - 1].description = value;\n      this.save();\n    }\n  \n    completed(status, index) {\n      this.todos[index - 1].completed = status;\n      this.save();\n    }\n  \n    clearCompleted() {\n      this.todos = this.todos.filter((a) => a.completed === false);\n      for (let i = 0; i < this.todos.length; i += 1) {\n        this.todos[i].index = i + 1;\n      }\n      this.save();\n    }\n  }\n\n//# sourceURL=webpack://webpack-template/./src/modules/classList.js?");
+
+/***/ }),
+
+/***/ "./src/modules/clear.js":
+/*!******************************!*\
+  !*** ./src/modules/clear.js ***!
+  \******************************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
+
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"clearCompleted\": () => (/* binding */ clearCompleted)\n/* harmony export */ });\n/* harmony import */ var _showList_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./showList.js */ \"./src/modules/showList.js\");\n\n\nconst clearCompletedBtn = document.getElementById('clear-completed');\n\nconst clearCompleted = () => {\n  clearCompletedBtn.addEventListener('click', () => {\n    _showList_js__WEBPACK_IMPORTED_MODULE_0__.todo.clearCompleted();\n    (0,_showList_js__WEBPACK_IMPORTED_MODULE_0__.create)();\n  });\n};\n\n//# sourceURL=webpack://webpack-template/./src/modules/clear.js?");
+
+/***/ }),
+
+/***/ "./src/modules/populate.js":
+/*!*********************************!*\
+  !*** ./src/modules/populate.js ***!
+  \*********************************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
+
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"populate\": () => (/* binding */ populate)\n/* harmony export */ });\n/* harmony import */ var _showList_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./showList.js */ \"./src/modules/showList.js\");\n\n\nconst populate = () => {\n    if (localStorage.getItem('todos')) {\n      _showList_js__WEBPACK_IMPORTED_MODULE_0__.todo.getReserved();\n      (0,_showList_js__WEBPACK_IMPORTED_MODULE_0__.create)();\n    } else {\n      (0,_showList_js__WEBPACK_IMPORTED_MODULE_0__.create)();\n    }\n};\n\n//# sourceURL=webpack://webpack-template/./src/modules/populate.js?");
+
+/***/ }),
+
+/***/ "./src/modules/showList.js":
+/*!*********************************!*\
+  !*** ./src/modules/showList.js ***!
+  \*********************************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
+
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"todo\": () => (/* binding */ todo),\n/* harmony export */   \"create\": () => (/* binding */ create)\n/* harmony export */ });\n/* harmony import */ var _classList_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./classList.js */ \"./src/modules/classList.js\");\n\n\nconst todo = new _classList_js__WEBPACK_IMPORTED_MODULE_0__.TaskList();\nconst listSection = document.querySelector('.list-section');\n\nconst create = () => {\n  listSection.replaceChildren();\n  if (todo.todos.length > 0) {\n    listSection.style.display = 'block';\n    const listContainer = document.createElement('ul');\n    listContainer.className = 'allTodos';\n    listSection.appendChild(listContainer);\n    todo.todos.map((a) => {\n      const list = document.createElement('li');\n      list.className = 'todo';\n\n      const descriptionContainer = document.createElement('div');\n      const checkbox = document.createElement('input');\n      checkbox.type = 'checkbox';\n      checkbox.id = 'checkbox';\n      if (a.completed === true) {\n        checkbox.checked = 'checked';\n      }\n\n      checkbox.onclick = (e) => {\n        todo.completed(e.target.checked, a.index);\n      };\n\n      descriptionContainer.appendChild(checkbox);\n\n      const descript = document.createElement('p');\n      descript.id = 'task-description';\n      descript.textContent = a.description;\n      descriptionContainer.appendChild(descript);\n      list.appendChild(descriptionContainer);\n\n      const editIcon = document.createElement('i');\n      editIcon.className = 'fa fa-solid fa-pen-to-square';\n      list.appendChild(editIcon);\n\n      const deleteIcon = document.createElement('i');\n      deleteIcon.className = 'fa fa-solid fa-trash';\n      deleteIcon.id = a.index;\n\n      list.onclick = () => {\n        descript.contentEditable = 'true';\n        list.style.backgroundColor = 'whitesmoke';\n        list.appendChild(deleteIcon);\n        editIcon.style.display = 'none';\n        list.addEventListener('keydown', () => {\n          todo.edit(descript.innerHTML, a.index);\n        });\n      };\n      deleteIcon.onclick = () => {\n        todo.delete(a.index);\n        todo.save();\n        create();\n      };\n      listContainer.append(list);\n      return list;\n    });\n    listSection.appendChild(listContainer);\n  }\n};\n\n//# sourceURL=webpack://webpack-template/./src/modules/showList.js?");
 
 /***/ })
