+++ src/modules/clear.js
@@ -1,0 +1,10 @@
+import { todo, create } from './showList.js';
+
+const clearCompletedBtn = document.getElementById('clear-completed');
+
+export const clearCompleted = () => {
+  clearCompletedBtn.addEventListener('click', () => {
+    todo.clearCompleted();
+    create();
+  });
+};
